build: compile with -Wextra
This commit also contains a few fixes to address those errors.
Functionality (except in extreme cases, i.e., greater than 0xf0000000
instructions) is unchanged.

Overruns are silently ignored. A proper fix would be to dynamically
allocate the input buffer, but at least this won't overwrite any memory.

+++ src/asm.c
@@ -60,7 +60,7 @@ asm_unit_write(FILE *restrict o, const struct asm_unit *restrict u)
 	struct ctf_ins *c;
 	char *pad = ""; // a singe 0 byte.
 
-	for (int i = 0; i < u->len; i++) {
+	for (size_t i = 0; i < u->len; i++) {
 		x = &u->ins[i];
 		switch (x->type) {
 		case I_GEN:
